car/requirements.txt
obd==0.7.2
paho-mqtt==1.6.1
PyYAML==6.0.1
